# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Control Backend (Gateway)
  version: "1.0.0"
  description: >
    REST API, проходящее через API Gateway. Все ответы завернуты в объект с полем `success`.
servers:
  - url: http://localhost:8000
    description: Local via docker-compose
paths:
...
  /v1/orders:
    get:
      tags: [Orders]
      summary: List current user's orders (admin can filter by userId)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
            enum: [created, in_progress, done, cancelled]
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [createdAt, total]
        - in: query
          name: sortDir
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: userId
          schema:
            type: string
          description: Only for admin
      responses:
        "200":
          description: Paged orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
    post:
      tags: [Orders]
      summary: Create order for current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: Validation error or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Get single order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [created, in_progress, done, cancelled]
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItem:
      type: object
      required: [product, quantity]
      properties:
        product:
          type: string
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          format: float
          minimum: 0
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status:
          type: string
          enum: [created, in_progress, done, cancelled]
        total:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Order'
    OrderListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Order'
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string